Move ISymbolExtensions to OmniSharp.Roslyn
