+++ src/OmniSharp.DotNetTest/TestFeaturesDiscover.cs
@@ -4,8 +4,8 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using OmniSharp.Abstractions.Services;
 using OmniSharp.DotNetTest.TestFrameworks;
-using OmniSharp.Extensions;
 using OmniSharp.Models;
+using OmniSharp.Roslyn.Utilities;
 
 namespace OmniSharp.DotNetTest
 {
